refactor netlink-sys
This commit introduces a bunch of breaking changes, but I think they
bring worthy improvements codewise.

1. merge `mio_socket` feature into `tokio_socket`

The `mio_socket` is tightly coupled to the `tokio_socket` feature and
adds complexity, while the use case seems very nice: most users that
go for mio will use the tokio stack, so they'd rather go for the tokio
socket directly.

2. split `netlink_sys::sys` in two separate module for the socket and
   the address implementation

3. always re-export `netlink_sys::socket::Socket`, no matter the
   feature. To allow this, the tokio socket is renamed `TokioSocket`
   and the smol socket is renamed `SmolSocket`

4. remove the `Protocol` enum. It provided a little bit of type
   safety, but feel overkill since only ever use it to open a socket
   and never match against it for example.

5. cleanup the constants module. Move most of them to where they
   belong: the `netlink_packet_route` crate

6. finally, start addressing [issue 107] by starting documenting
   `netlink_sys` properly

[issue 107]: https://github.com/little-dude/netlink/issues/107
