+++ audit/src/lib.rs
@@ -1,27 +1,25 @@
-#![allow(clippy::module_inception)]
-
-use netlink_proto::{sys::Protocol, Connection};
-
 mod handle;
 pub use crate::handle::*;
 
 mod errors;
 pub use crate::errors::*;
 
-use futures::channel::mpsc::UnboundedReceiver;
 pub use netlink_packet_audit as packet;
+pub mod proto {
+    pub use netlink_proto::{Connection, ConnectionHandle, Error, ErrorKind};
+}
 pub use netlink_proto::sys;
 
-use packet::{AuditMessage, NetlinkMessage};
 use std::io;
-use sys::SocketAddr;
+
+use futures::channel::mpsc::UnboundedReceiver;
 
 #[allow(clippy::type_complexity)]
 pub fn new_connection() -> io::Result<(
-    Connection<AuditMessage>,
+    proto::Connection<packet::AuditMessage>,
     Handle,
-    UnboundedReceiver<(NetlinkMessage<AuditMessage>, SocketAddr)>,
+    UnboundedReceiver<(packet::NetlinkMessage<packet::AuditMessage>, sys::SocketAddr)>,
 )> {
-    let (conn, handle, messages) = netlink_proto::new_connection(Protocol::Audit)?;
+    let (conn, handle, messages) = netlink_proto::new_connection(sys::protocols::NETLINK_AUDIT)?;
     Ok((conn, Handle::new(handle), messages))
 }
