+++ netlink-proto/src/handle.rs
@@ -4,8 +4,8 @@ use netlink_packet_core::NetlinkMessage;
 use std::fmt::Debug;
 
 use crate::errors::{Error, ErrorKind};
+use crate::sys::SocketAddr;
 use crate::Request;
-use netlink_sys::SocketAddr;
 
 /// A handle to pass requests to a [`Connection`](struct.Connection.html).
 #[derive(Clone, Debug)]
