+++ netlink-sys/Cargo.toml
@@ -36,8 +36,7 @@ version = "1.1"
 
 [features]
 default = []
-mio_socket = ["mio"]
-tokio_socket = ["mio_socket", "tokio", "futures"]
+tokio_socket = ["mio", "tokio", "futures"]
 smol_socket = ["async-io"]
