+++ rtnetlink/Cargo.toml
@@ -26,4 +26,3 @@ netlink-proto = { path = "../netlink-proto", version = "0.4" }
 env_logger = "0.7.1"
 ipnetwork = "0.15.1"
 tokio = { version = "0.2.6", features = ["macros", "rt-core"] }
-netlink-sys = {path = "../netlink-sys", version = "0.4" }
