+++ rtnetlink/src/connection.rs
@@ -1,13 +1,12 @@
-use futures::channel::mpsc::UnboundedReceiver;
 use std::io;
+
+use futures::channel::mpsc::UnboundedReceiver;
 
 use crate::{
     packet::{NetlinkMessage, RtnlMessage},
+    proto::Connection,
+    sys::{protocols::NETLINK_ROUTE, SocketAddr},
     Handle,
-};
-use netlink_proto::{
-    sys::{Protocol, SocketAddr},
-    Connection,
 };
 
 #[allow(clippy::type_complexity)]
@@ -16,6 +15,6 @@ pub fn new_connection() -> io::Result<(
     Handle,
     UnboundedReceiver<(NetlinkMessage<RtnlMessage>, SocketAddr)>,
 )> {
-    let (conn, handle, messages) = netlink_proto::new_connection(Protocol::Route)?;
+    let (conn, handle, messages) = netlink_proto::new_connection(NETLINK_ROUTE)?;
     Ok((conn, Handle::new(handle), messages))
 }
